+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - added support for Solidity, Crystal and Haskel languages in semantic search
 - added some worktree related functions into the Extensions API
 - clarify missing API key warning when using local inference via Aider
+- added proxy support to when downloading dependencies
 
 ## [0.71.0]